requirements.txt
torch
torchvision
matplotlib
tqdm
numpy
discord
pillow
gunicorn

